+++ src/model/evidence/reads.rs
@@ -152,7 +152,7 @@ pub const EDIT_BAND: usize = 2;
 pub struct IndelEvidence {
     gap_params: IndelGapParams,
     pairhmm: pairhmm::PairHMM,
-    window: u32,
+    max_window: u32,
 }
 
 impl IndelEvidence {
@@ -162,7 +162,7 @@ impl IndelEvidence {
         prob_deletion_artifact: LogProb,
         prob_insertion_extend_artifact: LogProb,
         prob_deletion_extend_artifact: LogProb,
-        window: u32,
+        max_window: u32,
     ) -> Self {
         IndelEvidence {
             gap_params: IndelGapParams {
@@ -172,7 +172,7 @@ impl IndelEvidence {
                 prob_deletion_extend_artifact: prob_deletion_extend_artifact,
             },
             pairhmm: pairhmm::PairHMM::new(),
-            window,
+            max_window,
         }
     }
 
@@ -231,27 +231,30 @@ impl AbstractReadEvidence for IndelEvidence {
                 (Some(qstart), Some(qend)) => {
                     let qstart = qstart as usize;
                     let qend = qend as usize;
-                    let read_offset = qstart.saturating_sub(self.window as usize);
-                    let read_end = cmp::min(qend + self.window as usize, read_seq.len());
+                    // ensure that distance between qstart and qend does not make the window too
+                    // large
+                    let max_window = (self.max_window as usize).saturating_sub(qend - qstart);
+                    let read_offset = qstart.saturating_sub(max_window);
+                    let read_end = cmp::min(qend + max_window as usize, read_seq.len());
                     (read_offset, read_end, varstart as usize, true)
                 }
 
                 (Some(qstart), None) => {
                     let qstart = qstart as usize;
-                    let read_offset = qstart.saturating_sub(self.window as usize);
-                    let read_end = cmp::min(qstart + self.window as usize, read_seq.len());
+                    let read_offset = qstart.saturating_sub(self.max_window as usize);
+                    let read_end = cmp::min(qstart + self.max_window as usize, read_seq.len());
                     (read_offset, read_end, varstart as usize, true)
                 }
                 (None, Some(qend)) => {
                     let qend = qend as usize;
-                    let read_offset = qend.saturating_sub(self.window as usize);
-                    let read_end = cmp::min(qend + self.window as usize, read_seq.len());
+                    let read_offset = qend.saturating_sub(self.max_window as usize);
+                    let read_end = cmp::min(qend + self.max_window as usize, read_seq.len());
                     (read_offset, read_end, varend as usize, true)
                 }
                 (None, None) => {
                     let m = read_seq.len() / 2;
-                    let read_offset = m.saturating_sub(self.window as usize);
-                    let read_end = cmp::min(m + self.window as usize, read_seq.len());
+                    let read_offset = m.saturating_sub(self.max_window as usize);
+                    let read_end = cmp::min(m + self.max_window as usize, read_seq.len());
                     let breakpoint = record.pos() as usize + m;
                     // The following should only happen with deletions.
                     // It occurs if the read comes from ref allele and is mapped within start
@@ -266,7 +269,7 @@ impl AbstractReadEvidence for IndelEvidence {
         let start = start as usize;
         // the window on the reference should be a bit larger to allow some flexibility with close
         // indels. But it should not be so large that the read can align outside of the breakpoint.
-        let ref_window = (self.window as f64 * 1.5) as usize;
+        let ref_window = (self.max_window as f64 * 1.5) as usize;
 
         // read emission
         let read_emission = ReadEmission::new(&read_seq, read_qual, read_offset, read_end);
@@ -327,7 +330,7 @@ impl AbstractReadEvidence for IndelEvidence {
         assert!(!prob_ref.is_nan());
         assert!(!prob_alt.is_nan());
 
-        // Normalize probabilities. By this, we avoid biases due to proximal variants that are in
+        // METHOD: Normalize probabilities. By this, we avoid biases due to proximal variants that are in
         // cis with the considered one. They are normalized away since they affect both ref and alt.
         // In a sense, this assumes that the two considered alleles are the only possible ones.
         // However, if the read actually comes from a third allele, both probabilities will be
@@ -336,12 +339,19 @@ impl AbstractReadEvidence for IndelEvidence {
         // probabilities is relevant!
 
         if prob_ref != LogProb::ln_zero() && prob_alt != LogProb::ln_zero() {
-            // Only perform normalization if both probs are non-zero
+            // METHOD: Only perform normalization if both probs are non-zero
             // otherwise, we would artificially magnify the ratio
             // (compared to an epsilon for the zero case).
             let prob_total = prob_alt.ln_add_exp(prob_ref);
             prob_ref -= prob_total;
             prob_alt -= prob_total;
+        }
+
+        if prob_ref == LogProb::ln_zero() && prob_alt == LogProb::ln_zero() {
+            // METHOD: if both are zero, use 0.5 instead. Since only the ratio matters, this
+            // has the same effect, without rendering the entire pileup likelihood zero.
+            prob_ref = LogProb::from(Prob(0.5));
+            prob_alt = prob_ref;
         }
 
         Ok(Some((prob_ref, prob_alt)))